requirements.txt
asgiref==3.9.1
crispy-bootstrap5==2025.6
Django==5.2.6
django-crispy-forms==2.4
sqlparse==0.5.3
tzdata==2025.2
